requirements.txt
CacheControl==0.12.10
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.10
click==8.0.3
firebase-admin==5.2.0
Flask==2.0.2
google-api-core==2.4.0
google-api-python-client==2.36.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-cloud-vision==2.6.3
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.3
packaging==21.3
proto-plus==1.19.8
protobuf==3.19.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
requests==2.27.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2
